Product Admin and Profile Admin views secured

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models.functions import Lower
 # Used to handle complex query in the database to achieve 'or' logic
 # To match either from product name or product description - use Q
@@ -75,8 +76,13 @@ def product_detail(request, product_id):
     return render(request, 'products/product_detail.html', context)
 
 
+# Decorator to wrap around the function to check user access before executing
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
 
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
@@ -97,8 +103,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+# Decorator to wrap around the function to check user access before executing
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -121,8 +133,14 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+# Decorator to wrap around the function to check user access before executing
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
